requirements.txt
bluepy
pycrypto
requests
